fixed email reporter

+++ run_net_control.sh
@@ -9,6 +9,7 @@ export SERVER_DIR=$(cat server_dir)
 export SERVER_URL=$(cat server_url)
 export USER_MAIL=$(cat useremail)
 
+NL=$'\n'
 
 export MOKSISKAAN_HOME=/opt/moksiskaan/db
 #export EXEC_DIR=/home/vrogojin/biomedicum/FBL-paper-exec
@@ -25,7 +26,9 @@ ssh -i ~/.ssh/id_rsa -o 'StrictHostKeyChecking no' -o 'UserKnownHostsFile /dev/n
 ../../netco4biomed.and &
 ID=$!
 
-../../notification_sender.sh postmaster@combio.abo.fi $USER_MAIL "TASK $SESSION_ID STARTED" "Hello,\\nYou have started task at NetControl4BioMed pipeline.\\nFollow the progress of the analysis at http://$SERVER_URL/$SESSION_ID/result.html" 2>&1 > mail.log
+
+../../notification_sender.sh postmaster@combio.abo.fi $USER_MAIL "TASK $SESSION_ID STARTED" \
+    "Hello,${NL}You have started task at NetControl4BioMed pipeline.${NL}Follow the progress of the analysis at $SERVER_URL/$SESSION_ID/result.html" 2>&1 > mail.log
 
 echo "<html>" > pipeline.log.html
 echo "<body>" >> pipeline.log.html
@@ -39,6 +42,8 @@ echo "</html>" >> pipeline.log.html
 echo "<font color=orange>PIPELINE RUNNING...</font>" > status.txt
 ../../generate_result_page.sh
 scp -i ~/.ssh/id_rsa -o 'StrictHostKeyChecking no' -o 'UserKnownHostsFile /dev/null' pipeline.log.html result.html frontend:$SERVER_DIR
+
+result_not_ready=1
 
 running=1
 while [ $running -eq 1 ]; do
@@ -82,9 +87,14 @@ while [ $running -eq 1 ]; do
     then
 	scp -i ~/.ssh/id_rsa -o 'StrictHostKeyChecking no' -o 'UserKnownHostsFile /dev/null' result/network.graphml frontend:$SERVER_DIR/result
     fi
-    if [ -e "result.zip" ]
+    if [ -s "result.zip" ]
     then
 	scp -i ~/.ssh/id_rsa -o 'StrictHostKeyChecking no' -o 'UserKnownHostsFile /dev/null' result.zip frontend:$SERVER_DIR
+	if [ $result_not_ready -eq 1 ]; then
+	    ../../notification_sender.sh postmaster@combio.abo.fi $USER_MAIL "TASK $SESSION_ID ALMOST COMPLETED." \
+		"Hello,${NL}The results of your network control analysis are ready and available for download.${NL}The PDF plot with the network is being generated. Depending of the size of the network, this may take from a couple of minutes to many hours.${NL}If the network is too large, PDF generation may fail.${NL}For your convenience, visit $SERVER_URL/$SESSION_ID/result.html" 2>&1 > mail.log
+	    result_not_ready=0
+	fi
     fi
     if grep -q 'STOP' control.txt; then
 	pkill --signal 9 -P "$ID"
@@ -107,15 +117,18 @@ echo "$ID TERMINATED" >> out.txt
 cp log_netco4biomed/_global pipeline.log
 #cat ../../result_page_template_header.html | sed "s/###/$SESSION_ID/g" | sed "s/<meta http-Equiv=\"Refresh\" Content=\"5\">//g" > result.html
 
-
+ST="RUNNING"
 if [ -s res ]
 then
     echo "<font color=green>PIPELINE TERMINATRED. STATUS OK</font>" > status.txt
+    ST="OK"
 else
     if grep -q 'STOP' control.txt; then
 	echo "<font color=red>PIPELINE INTERRUPTED BY USER. CHECK THE LOG</font>" > status.txt
+	export ST="INTERRUPTED"
     else
 	echo "<font color=red>PIPELINE FAILED. CHECK THE LOG</font>" > status.txt
+	export ST="FAILED"
     fi
 fi
 ../../generate_result_page.sh
@@ -129,3 +142,5 @@ scp -i ~/.ssh/id_rsa -o 'StrictHostKeyChecking no' -o 'UserKnownHostsFile /dev/n
 #emailcommand="ssh -i ~/.ssh/id_rsa -o 'StrictHostKeyChecking no' -o 'UserKnownHostsFile /dev/null' frontent $emailcommand"
 #eval '$emailcommand'
 
+../../notification_sender.sh postmaster@combio.abo.fi $USER_MAIL "TASK $SESSION_ID COMPLETED. STATUS $ST" \
+    "Hello,${NL}Your data analysis session completed.${NL}Status: ${ST}.${NL}For your convenience, visit $SERVER_URL/$SESSION_ID/result.html" 2>&1 > mail.log
